services/ingest_service/requirements.txt
python-dotenv>=1.0.1
fastapi>=0.109.2
pydantic<=2.6.1
openai>=1.12.0
llama-index>=0.9.47
uvicorn[standard]>=0.27.1
firebase-admin>=6.4.0
unstructured>=0.12.4
pinecone-client>=3.0.2
google-cloud-storage>=2.14.0
